refactor: constants module into a single file
