--- src/constants/config_constants.py
@@ -1,6 +1,0 @@
-# Average distance between human eyes is 63 mm. 
-# This is 6.3 cm. 
-INTEROCULAR_DISTANCE = 6.3  
-
-# Focal length of the camera in use
-FOCAL_LENGTH = 655
