+++ src/estimate_distance.py
@@ -1,5 +1,5 @@
-from constants.model_constants import CVZONE_DETECTOR_MAX_ONE, EYE_DISTANCE_INDEX
-from constants.config_constants import FOCAL_LENGTH, INTEROCULAR_DISTANCE
+from constants import CVZONE_DETECTOR_MAX_ONE, EYE_DISTANCE_INDEX
+from config_constants import FOCAL_LENGTH, INTEROCULAR_DISTANCE
 
 
 def estimate_depth(landmarks):
